JCHS section
Added a section of the website called JCHS.  The idea is to house the outward facing content for JCHS there.  However, most of the content will probably be moved out of this and into a seperate repository.  Divided up based on content area.

It is possible that they will be merged back in, but for now I am much more comfortable formatting my thoughts into PDFs using LaTeX.